DevOacc: Add for same iDevice return same impl instance every time
